LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-createHTMLDocument-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
shouldpass [default]

FAIL Trusted Type assignment is blocked. (document) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
PASS Trusted Type instances created in the main doc can be used. (document)
FAIL Trusted Type assignment is blocked. (createHTMLDocument) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
PASS Trusted Type instances created in the main doc can be used. (createHTMLDocument)
FAIL Trusted Type assignment is blocked. (DOMParser) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
PASS Trusted Type instances created in the main doc can be used. (DOMParser)
FAIL Trusted Type assignment is blocked. (XHR) assert_throws_js: function "_ => { doc.createElement("script").textContent = "2+2"; }" did not throw
PASS Trusted Type instances created in the main doc can be used. (XHR)
PASS Install default policy.
PASS Default policy applies. (document)
PASS Default policy applies. (createHTMLDocument)
PASS Default policy applies. (DOMParser)
PASS Default policy applies. (XHR)
